+++ js/pages/cohort-definitions/components/checks/conceptset-warnings.js
@@ -1,5 +1,5 @@
 define(['knockout', 'text!./conceptset-warnings.html',
-    'services/CohortDefinition',
+    'services/CohortDefinitionService',
     './const',
     './utils',
     'databindings',
