+++ js/pages/cohort-definitions/components/reporting/cost-utilization/base-drug-util-report.js
@@ -7,7 +7,7 @@ define(
     'utils/BemHelper',
     'appConfig',
     'pages/cohort-definitions/const',
-    'services/CohortResults',
+    'services/CohortResultsService',
     'components/visualizations/filter-panel/filter-panel',
     'less!./drug-util-detailed.less',
   ],
