+++ js/pages/cohort-definitions/components/reporting/cost-utilization/visit-util.js
@@ -4,7 +4,7 @@ define(
     'text!./visit-util.html',
     'pages/cohort-definitions/components/reporting/cost-utilization/base-report',
     'pages/cohort-definitions/const',
-    'services/CohortResults',
+    'services/CohortResultsService',
     'appConfig',
     'utils/CommonUtils',
     'components/visualizations/filter-panel/filter-panel',
