+++ js/providers/Service.js
@@ -1,0 +1,13 @@
+define([
+	'services/http',
+], (
+	httpService,
+) => {
+
+	return class Service {
+		constructor() {
+			this.httpService = httpService;
+		}		
+	}
+
+});
